Add individual ticket setting for VoteBits via the RPC
This update allows the per-ticket setting of the VoteBits field so
that different VoteBits settings can be user selected in stake pools.
It is backwards compatible with the former VoteBits selection method
on the command line. Tickets stored before this update will continue
to use the VoteBits selection from the command line. Tickets stored
after this update will be initialized with the VoteBits selection from
the command line and then be able to be updated by the RPC. After
updating, any votes cast with that ticket will be assigned the
per-ticket, user selected VoteBits instead of the default VoteBits.