Disable unsafe RPC calls on MainNet and fix extended pubkey storage
RPC commands that allow the removal of unencrypted private data from the
wallet (dumpprivkey, getseed) have been disabled on MainNet wallets. The
storage of the seed for a MainNet wallet has also been disabled. These
rules can be bypassed with the use of the new '--unsafemainnet' flag. The
storage and retrieval of the master extended public key used to generate a
watching only wallet has been refactored and made safer.