Fix bug in multisignature output spending in chainntfns.go
A bug in multisignature handling would occasionally cause the
inappropriate failing of transactions being added to the transaction
manager. If a multisignature script of any kind was found but a
reference to the script could not be found in the transaction
manager, the whole function would fail. The error returned from
the function is now ignored because failing to find a multisignature
output to spend for a random multisignature script should not be
considered an error.