Fix a bug that caused votes and revocations not being stored
In the case of redundant wallets, votes and revocations that were
rejected from the daemon as duplicates were not stored at all in
the stake database. This would cause corrupted values for votes
and revocations when calling getstakeinfo. This was corrected by
storing the votes and revocations first, before shipping to the
daemon.