Fix a bug relating to relevantTx handling and uncaught error
A call to UnspentOutpoints did not have error handling, which was
added.

A restrictive switch statement would cause all outputs that were not
either pay-to-pubkey hash or pay-to-script-hash multisig to be stored.
This was reverted to the same behavior as btcwallet where all outputs
are stored if the address is found in the address manager. Code
handling multisignature redeemScript lookup in outputs was also
corrected.