Refactor address pool code and automatically resync accounts from seed
The old address pool code was faulty in that it could not scale to
more accounts than the default account and did address index alignment
using an actual address instead of an index. This updates the address
pools to be generalized so that in the future they can be multiaccount.
Address indexes for an account are now stored in the meta bucket of the
address manager database. The logic for the address pool has been
greatly simplified.

Accounts are now restored when the wallet is restored from seed. New
logic similar to the logarithmic scan for addresses in the default
account has been added. On start up, the wallet will also rescan and
check for new address usage in all of its used accounts. On first start
up, the wallet must be unlocked so that accounts can be added if
necessary. New logic prompts the user for the password.

A new flag, --promptpass, has been added. This will cause the wallet
to prompt the user for the password on startup, leaving the wallet
unlocked indefinitely.