Ticket purchasing code overhaul (#183)
* Ticket purchasing code overhaul

This pull request removes the old, buggy ticket purchase code and
replaces it with code that is more robust. In particular, chains of
tickets can now be purchased with ease because outputs to be consumed
by tickets are now pregenerated in split transactions. Ticket expiry
can now be added to ensure that the tickets will expire in a short time
frame, so that the end user can update the ticket fees if their tickets
fail to be added to a block.

A new feature that was added was pool fees. Tickets can now pay a
fee to a pool that is set by the user. The fee determines the amount
that the user will contribute to the pool. The pool then receives this
amount plus subsidy.