Open and return wallet from CreateNewWallet.
This fixes a dcrwallet-specific bug introduced by
4b64adf52f873a4ec5237d3fb16b5252ddef3846 which removed the call to
wallet.Open from wallet.Loader.CreateNewWallet, replacing the wallet
pointer with nil, and removes the addition of a call to close the
wallet database.  The introduction of a nil pointer caused runtime
panics when executing Loader callbacks that dereferenced the wallet
pointer, and by the caller who would dereference the return value.

This makes the LoaderService.CreateNewWallet RPC useful again.