Fix bugs relating to reorganizations (#236)
Two bugs relating to reorganizations were fixed. The first was an off
by one bug on initial sync which rarely would ever affect the end user.
The second was that logic causing the removal of all blocks to the
genesis block on a disconnection of an unknown block would cause
irreversible corruption of the wallet. The new code ignores the hash
of the block to remove and assumes that daemon removing blocks to some
common unknown forking point between the daemon and wallet before
connecting the side chain's blocks.