Change n in RangeTransactions to more closely track count param (#280)
When using listtransactions it was not incrementing tx correctly, especially when there were transactions in mempool.  Now when listtransactions count option is given, it will return all the vin/vouts for a given txid.  So when you look at the results of listtransactions you may see more entries than the count you offered, but you should see the equal number of unique txids