Improve wallet atomicity. (#339)
This is a step towards completing #315.

This changes the database access APIs and each of the "manager"
packages (waddrmgr/wtxmgr/wstakemgr) so that transactions are
opened (only) by the wallet package and the namespace buckets
that each manager expects to operate on are passed in as
parameters.

This helps improve the atomicity situation as it means that many
calls to these APIs can be grouped together into a single
database transaction.

This change does not attempt to completely fix
the "half-processed" block problem. Mined transactions are still
added to the wallet database under their own database transaction
as this is how they are notified by the consensus JSON-RPC
server (as loose transactions, without the rest of the block that
contains them). It will make updating to a fixed notification
model significantly easier, as the same "manager" APIs can still
be used, but grouped into a single atomic transaction.