Allow running either the new or old ticket buyer. (#470)
An --enableticketbuyer option has been added.  It must be used to
enable the new ticket buyer.

An --enablevoting option has been added to turn on the wallet's
automatic creation of vote and revocation transactions when using the
new ticket purchaser.

The --enablestakemining option has been deprecated for the above two
options.  To use the old ticket buyer from the wallet package, use
this option and do NOT set --enableticketbuyer (doing so would enable
the new ticket buyer).

All RPCs to interact with the ticket buyer remain, but only affect the
old deprecated ticket buyer, and are also deprecated by extension.

Fixes #468.