Add method to retrieve an account branch xpub.
This new method takes a read-only database transaction, rather than
the bucket.  This differs from the rest of the API which currently
takes namespace buckets themselves, but this is the direction that I
want to take the API.

To remove some generated garbage created by creating the bucket key
bytes, do the conversion once and save the keys as package variables.
Use these consistently across the entire package.