Fix an off by one in the address buffer.
This fixes two issues:

1. The old code would always hand out new addresses starting at the
   discovered last used child index.  This could result in accidental
   address reuse after restarting the wallet.

2. When the address gap limit was exhausted and the cursor wrapped
   around to the beginning, the next returned address would skip over
   the first returned address, because it was used.

To fix this, there needed to be a way to mark no addresses as used,
instead of recording child 0 as used.  This commit makes a change to
the database upgrade code and records the last used child index
^uint32(0) in this case.  This had to be a change to previous update
code as there is no way to otherwise distinguish between no use at all
or usage of just the 0th child after the upgrade has completed.