Allow reading passphases from piped input.
This change checks if stdin is a tty or not, and when it is not, it just reads
the input from stdin without removing terminal echo.  This allows passphrases to
be piped to the process.

For users running into difficulty starting dcrwallet from startup scripts or as
a service due to the passphrase prompt when enabling the ticket buyer, it is now
possible to use the pipe as a workaround:

  dcrwallet --enableticketbuyer < passfile
  echo $pass | dcrwallet --enableticketbuyer