Raise default fees back to 0.01/kB.
Users can still lower the fees on their own but until the majority of
the network has upgraded to the new mempool policy defaults, lowering
these fees can cause transactions to be rejected and votes to be
missed.

While here, remove the hardcoded constant used for revocations so
these are also created using the wallet's current relay fee.