Remove WalletService.SpentnessNotification RPC.
This RPC has never worked and shouldn't have been added to the API
because it was unimplemented.

I'm not bumping the API major version in this case, even though
technically it is a breaking change, because the API has never worked
and we know there are no callers who are relying on it.