Rollback DB updates for failed tx publishing.
This change improves the transaction creation error handling to handle
failures for rejected transactions or errors storing a transaction
into the database.  When purchasing tickets (a task that requires
creating multiple transactions), a DB update is now opened for each
published transaction.

The code has also been improved to only open DB updates when storing
or publishing the transaction.  Views are used when authoring and
signing transactions.  This reduces the amount of time spent in the
update and should improve the performance characteristics of the
wallet.

In the future, transactions could be published outside of the update,
but this will require a second update to remove any stored
transactions if the send fails.  This can not be done at this time due
to missing the udb APIs to remove these transactions.