Provide means to confirm imported scripts are redeemable.
The WalletService.ImportScript RPC request message has gained a new
field to indicate that the RPC should error out early and not import
the script if the wallet does not have enough of the required secrets
to redeem a multisig script.  The response message has gained fields
for the P2SH address created from the imported script and a bool for
whether an imported script (where the script was not required to be
redeemable in the request) is known to be redeemable anyways.