Add wallet APIs for controlling the gap limit policy.
There are three policies defined:

1. Error when the unused address gap limit is violated

2. Ignore the gap limit and generate the next child address
   regardless

3. Wrap around back to the first unused address after the last used
   address to avoid violating the gap limit

The third wrapping policy was the default and only policy up until
now.  The new default is to error, with the intention of allowing
users to specify the policy they desire.

Because nothing calls the new address methods yet with anything other
than the wrapping policy, this commit introduces no functional change
to the program's behavior, but lays the groudwork for the policy to be
exposed to users elsewhere (such as options in RPC requests or the
wallet config).