Error for missing expicitly specified configs.
This uses the new cfgutil.ExplicitString to ensure that the option was
really explicitly set instead of comparing the value itself to a
default.  Unlike the implementation in 1020bdd, this works correctly
when the default config file location is changed due to a modified
appdata directory.

While here, update the rest of the config options that are relative to
the appdata directory to use ExplicitString.