Track all tickets and their spent status.
Prior to this change, all stake transactions were tracked in the
"stake manager", a set of buckets that recorded (duplicate) ticket
transactions, block metadata for tickets/votes/revocations, and which
tickets are owned by which stakepool users.  However, the stake
manager could not provide a consistent view of these transactions as
it was not correctly updated with changes to the blockchain.  For
example, it would provide incorrect information if a reorg occurred or
if votes were created but missed.

This change introduces ticket tracking into the transaction manager
and switches over several key components to use this tracking instead
of the hacked-on stake manager.  The advantage of integrating this
with all other block and transaction processing is that we get spend
tracking can be reused to determine whether a ticket was spent by a
vote or revocation.  Unlike the stake manager, the transaction manager
is aware of processed blocks and all other transactions and can
provide a consistent view of all tickets.

In the future, the tickets bucket will also record the picked status
of a ticket.  This will allow the wallet to know which block a ticket
was picked, and makes it possible to find all revokable tickets
without invoking additional dcrd RPCs.

The udb package now provides a type that can be used to iterate over
all tickets.  This iterator has been used to reimplement the
Wallet.StakeInfo method (improving getstakeinfo's performance by about
20-30%) and will be used in future RPCs to return the states of all
tickets as a stream.

The old stake manager is still kept and must remain until stakepools
no longer require it.  At this point in time, the stake manager is
still recording pool user info that is used by the stakepool frontend.

This change introduces a database upgrade and older versions of the
wallet will not work on upgraded wallet DBs.