txauthor: Update for MsgTx API changes.
This updates the txauthor package to set the new serialization type
field in the wire.MsgTx struct and introduces a local constant which
defines the version to create the transaction with instead of using the
latest version in wire.  This is important since a change in the
transaction version will potentially require changes to the generated
transaction and thus using the wire constant for the generated
transaction version could allow creation of invalid transactions for the
updated version.