Handle address wraping policy after restarts.
When called with the wrapping gap policy, RPCs like getnewaddress and
WalletService.NextAddress would wrap around to a previously returned
address in order to avoid violating the BIP0044 gap limit.  However,
this was not correctly handled on restarts due to the last returned
child indexes not being reset down to the lower values.  This changes
the database code so that lower returned child indexes are written and
persisted across restarts.