Register all gRPC services before serving.
It is not allowed to register additional gRPC services after the
server has started serving.  However, this was being done for optional
services that weren't ready to be used yet (e.g. the WalletServer
would be started later in the program's execution after being loaded).

This bug was discovered due to a check and fatal log in newer versions
of the Go gRPC implementation.

The fix used here is to create incomplete implementations of all
services and register each of them before serving.  The method
interceptors have added a check to verify that the service is ready
and available before the methods are called on the incomplete service.
Atomics are used to check for service readiness when necessary.  Like
before (where gRPC would panic when a service was registered multiple
times) the rpcserver package now panics if the service is readied
multiple times.