Allow JSON-RPC handlers access to the wallet loader.
This is now being passed into the walletinfo handler to provide
information on whether the ticket buyer is running.  This is
preferable to querying this from the wallet itself since the wallet no
longer purchases tickets and the ticket buyer can be started/stopped
at will (over gRPC) so plumbing the config option to the handler
through the wallet struct is not the correct way to report this.

Remove the old TicketPurchasingEnabled methods from the wallet as they
are simply no longer needed.