Fix previous database upgrade.
The previous database upgrade wrote data to the wrong bits and could
cause ticket outputs to be wrongly chosen as spendable outputs to use
for regular transactions.  This change introduces another database
upgrade which corrects the previous issue and updates comments
describing the format of saved values to make these mistakes less
likely to occur in the future.